Move createPassportHash and emailTemplate to shared file
