+++ modules/user/server-ts/index.ts
@@ -16,8 +16,6 @@ export interface ValidationErrors {
   password?: string;
 }
 
-export const createPasswordHash = (pswd: string) => bcrypt.hash(pswd, 12);
-
 const getIdentity = (id: number) => {
   return UserDAO.getUser(id);
 };
